Move tests out of testUtils
