+++ test/testUtils.cpp
@@ -73,112 +73,6 @@ namespace LiFFTTest {
         write2File(name, data, LiFFT::accessors::makeTransformAccessor(acc, LiFFT::policies::CalcIntensityFunc()));
     }
 
-    void
-    testView()
-    {
-        using Extents = LiFFT::types::Vec<2>;
-        const Extents size(100u, 120u);
-        const Extents offset(10u, 5u);
-        const Extents viewSize(40u, 65u);
-        auto data = LiFFT::mem::RealContainer<2, float>(size);
-        auto view = LiFFT::types::makeView(data, LiFFT::types::makeRange(offset, viewSize));
-        Extents idx, viewIdx;
-        bool error = false;
-        for(idx[0] = offset[0], viewIdx[0] = 0; viewIdx[0] < viewSize[0]; ++idx[0], ++viewIdx[0])
-            for(idx[1] = offset[1], viewIdx[1] = 0; viewIdx[1] < viewSize[1]; ++idx[1], ++viewIdx[1])
-            {
-                float value = rand() / RAND_MAX;
-                view(viewIdx) = value;
-                if(data(idx) != view(viewIdx) || data(idx) != value)
-                    error = true;
-            }
-        if(error)
-            std::cerr << "View test failed" << std::endl;
-        else
-            std::cout << "View test passed" << std::endl;
-    }
-
-    unsigned getIdx(unsigned idx, unsigned fixedDim)
-    {
-        return idx >= fixedDim ? idx + 1 : idx;
-    }
-
-    template< unsigned fixedDim >
-    void
-    testSliceView()
-    {
-        using Extents3 = LiFFT::types::Vec<3>;
-        using Extents2 = LiFFT::types::Vec<2>;
-        const Extents3 size(100u, 120u, 214u);
-        const Extents3 offset(10u, 5u, 23u);
-        const Extents2 viewSize(40u, 65u);
-        auto data = LiFFT::mem::RealContainer<3, float>(size);
-        auto view = LiFFT::types::makeSliceView<fixedDim>(data, LiFFT::types::makeRange(offset, viewSize));
-        Extents3 idx;
-        Extents2 viewIdx;
-        bool error = false;
-        idx[fixedDim] = offset[fixedDim];
-        unsigned idxOut = getIdx(0, fixedDim);
-        unsigned idxIn = getIdx(1, fixedDim);
-        for(idx[idxOut] = offset[idxOut], viewIdx[0] = 0; viewIdx[0] < viewSize[0]; ++idx[idxOut], ++viewIdx[0])
-            for(idx[idxIn] = offset[idxIn], viewIdx[1] = 0; viewIdx[1] < viewSize[1]; ++idx[idxIn], ++viewIdx[1])
-            {
-                float value = rand() / RAND_MAX;
-                view(viewIdx) = value;
-                if(data(idx) != view(viewIdx) || data(idx) != value)
-                    error = true;
-            }
-        if(error)
-            std::cerr << "SliceView(" << fixedDim << ") test failed" << std::endl;
-        else
-            std::cout << "SliceView(" << fixedDim << ") test passed" << std::endl;
-    }
-
-    void
-    testDataWrappers()
-    {
-        const unsigned size = 100u;
-        using Extents = LiFFT::types::Vec<3>;
-        using FFT = LiFFT::FFT_3D_R2C_F<>;
-        auto input = FFT::wrapInput(
-                        LiFFT::mem::RealContainer<3, float>(
-                                Extents(size, size, size)
-                        )
-                     );
-        auto data = LiFFT::mem::RealContainer<3, float>(
-                Extents(size, size, size)
-        );
-        auto output = FFT::wrapInput(data);
-
-        Extents idx = Extents::all(0u);
-        const float val = 1337;
-        const float val2 = 1338;
-        bool error = false;
-        auto acc = LiFFT::traits::getIdentityAccessor(input);
-        auto acc2 = LiFFT::traits::getIdentityAccessor(output);
-        auto acc3 = LiFFT::traits::getIdentityAccessor(data);
-        for(unsigned i=0; i<4; i++){
-            input(idx) = val;
-            output(idx) = val;
-            if(input(idx) != val)
-                error = true;
-            if(output(idx) != val || acc3(idx, data) != val)
-                error = true;
-            acc(idx, input) = val2;
-            acc2(idx, output) = val2;
-            if(acc(idx, input) != val2 || input(idx)!=val2)
-                error = true;
-            if(acc2(idx, output) != val2 || acc3(idx, data) != val2 || output(idx)!=val2)
-                error = true;
-            if(i<3)
-                idx[i] = size/2;
-        }
-        if(error)
-            std::cerr << "Datawrapper test failed" << std::endl;
-        else
-            std::cout << "Datawrapper test passed" << std::endl;
-    }
-
     void init()
     {
         TestExtents size = TestExtents::all(testSize);
@@ -203,11 +97,6 @@ namespace LiFFTTest {
             fftC2C = static_cast<decltype(fftC2C)>(malloc(sizeof(FFT)));
             new(fftC2C)auto(FFT(input, output));
         }
-        testDataWrappers();
-        testView();
-        testSliceView<0>();
-        testSliceView<1>();
-        testSliceView<2>();
     }
 
     void finalize()
